[refact/#78]: 주소 API 코드 분리
