+++ components/common/index.js
@@ -1,4 +1,4 @@
-import { AddressPostcode } from './AddressPostcode';
+import { AddressFormField, AddressSearchModal } from './AddressPostcode';
 import { ButtonL, ButtonM, ButtonS } from './Button';
 import { CheckBox } from './CheckBox';
 import { Dropdown } from './Dropdown';
@@ -16,7 +16,8 @@ import { Title } from './Title';
 import { Toast } from './Toast';
 
 export {
-  AddressPostcode,
+  AddressFormField,
+  AddressSearchModal,
   ButtonL,
   ButtonM,
   ButtonS,
