+++ constants/address.js
@@ -1,0 +1,22 @@
+export const doList = [
+  '강원특별자치도',
+  '경기',
+  '경남',
+  '경북',
+  '전남',
+  '전북특별자치도',
+  '충남',
+  '충북',
+  '제주특별자치도',
+];
+
+export const majorCities = ['서울', '부산', '대구', '인천', '광주', '대전', '울산'];
+
+export const NONE = '없음';
+
+export const initEmpty = {
+  doName: NONE,
+  siName: NONE,
+  guName: NONE,
+  dongName: NONE,
+};
